+++ edx_proctoring/api.py
@@ -58,15 +58,17 @@ from edx_proctoring.serializers import (
 )
 from edx_proctoring.statuses import InstructorDashboardOnboardingAttemptStatus, ProctoredExamStudentAttemptStatus
 from edx_proctoring.utils import (
+    categorize_inaccessible_exams_by_date,
     emit_event,
     get_exam_due_date,
     get_exam_type,
+    get_exam_url,
     get_time_remaining_for_attempt,
+    get_user_course_outline_details,
     has_due_date_passed,
     humanized_time,
     is_reattempting_exam,
     obscured_user_id,
-    resolve_exam_url_for_learning_mfe,
     verify_and_add_wait_deadline
 )
 
@@ -79,6 +81,38 @@ APPROVED_STATUS = 'approved'
 REJECTED_GRADE_OVERRIDE_EARNED = 0.0
 
 USER_MODEL = get_user_model()
+
+
+def get_onboarding_exam_link(course_id, user, is_learning_mfe):
+    """
+    Get link to the onboarding exam available to the user for given course.
+
+    Parameters:
+    * course_id: id of a course where search for the onboarding exam is performed
+    * user: user for whom the link is built
+    * is_learning_mfe: Boolean, indicates if the url should be built for the learning mfe application.
+
+    Returns: url pointing to the available exam, if there is no onboarding exam available
+    to the user returns empty string.
+    """
+    onboarding_exams = list(ProctoredExam.get_practice_proctored_exams_for_course(course_id).order_by('-created'))
+    if not onboarding_exams or not get_backend_provider(name=onboarding_exams[0].backend).supports_onboarding:
+        onboarding_link = ''
+    else:
+        details = get_user_course_outline_details(user, course_id)
+        categorized_exams = categorize_inaccessible_exams_by_date(onboarding_exams, details)
+        non_date_inaccessible_exams = categorized_exams[0]
+
+        # remove onboarding exams not accessible to learners
+        for onboarding_exam in non_date_inaccessible_exams:
+            onboarding_exams.remove(onboarding_exam)
+
+        if onboarding_exams:
+            onboarding_exam = onboarding_exams[0]
+            onboarding_link = get_exam_url(course_id, onboarding_exam.content_id, is_learning_mfe)
+        else:
+            onboarding_link = ''
+    return onboarding_link
 
 
 def get_total_allowed_time_for_exam(exam, user_id):
@@ -752,10 +786,7 @@ def get_exam_attempt_data(exam_id, attempt_id, is_learning_mfe=False):
 
     # resolve the LMS url, note we can't assume we're running in
     # a same process as the LMS
-    if is_learning_mfe:
-        exam_url_path = resolve_exam_url_for_learning_mfe(exam['course_id'], exam['content_id'])
-    else:
-        exam_url_path = reverse('jump_to', args=[exam['course_id'], exam['content_id']])
+    exam_url_path = get_exam_url(exam['course_id'], exam['content_id'], is_learning_mfe)
 
     attempt_data = {
         'in_timed_exam': True,